feat: adiciona camada de segurança em ItemMenu

+++ src/main/java/br/com/fiap/TechChallenger/usecases/itemMenu/DeletarItemMenu.java
@@ -1,9 +1,13 @@
 package br.com.fiap.TechChallenger.usecases.itemMenu;
 
 import br.com.fiap.TechChallenger.domains.ItemMenu;
+import br.com.fiap.TechChallenger.domains.dto.UsuarioLogado;
 import br.com.fiap.TechChallenger.gateways.repository.ItemMenuRepository;
 import br.com.fiap.TechChallenger.usecases.exception.RecursoNaoEncontradoException;
+import br.com.fiap.TechChallenger.usecases.security.Autenticacao;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
@@ -12,15 +16,30 @@ import org.springframework.stereotype.Service;
 public class DeletarItemMenu {
 
     private final ItemMenuRepository itemMenuRepository;
+    private final Autenticacao autenticacao;
 
-    public ResponseEntity<Void> deletar(Long id) {
-        ItemMenu itemMenu = itemMenuRepository.findById(id)
-                .orElseThrow(() -> new RecursoNaoEncontradoException(
-                        "Id " + id + " não foi encontrado."
-                ));
+    public ResponseEntity<Void> deletar(Long id, HttpServletRequest request) {
 
-        itemMenuRepository.delete(itemMenu);
+        try {
+            UsuarioLogado usuarioLogado = autenticacao.getUsuarioLogado(request);
+            String perfil = usuarioLogado.getPerfil();
 
-        return ResponseEntity.noContent().build();
+            if (!"DONO_RESTAURANTE".equalsIgnoreCase(perfil)) {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
+
+            ItemMenu itemMenu = itemMenuRepository.findById(id)
+                    .orElseThrow(() -> new RecursoNaoEncontradoException(
+                            "Id " + id + " não foi encontrado."
+                    ));
+
+            itemMenuRepository.delete(itemMenu);
+
+            return ResponseEntity.noContent().build();
+
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
     }
 }
